Rename: `utils` -> `hooks`
